Created Message Popup + Book Photo Slideshow
Created the message pop-up on the userpage and booklisting page that can
be opened and closed by clicking outside the pop-up. Also created a
slideshow for book images on the booklisting page: clicking on each
picture in the navigation makes it the main image shown, and clicking
the main image makes the full-version of the image appear as a pop-up.